tests: switch from int to size_t for CURL_ARRAYSIZE()

+++ tests/libtest/lib1506.c
@@ -33,7 +33,7 @@ static CURLcode test_lib1506(char *URL)
   CURL *curl[NUM_HANDLES] = {0};
   int running;
   CURLM *m = NULL;
-  int i;
+  size_t i;
   char target_url[256];
   char dnsentry[256];
   struct curl_slist *slist = NULL, *slist2;
@@ -45,7 +45,7 @@ static CURLcode test_lib1506(char *URL)
   /* Create fake DNS entries for serverX.example.com for all handles */
   for(i = 0; i < CURL_ARRAYSIZE(curl); i++) {
     curl_msnprintf(dnsentry, sizeof(dnsentry), "server%d.example.com:%s:%s",
-                   i + 1, port, address);
+                   (int)i + 1, port, address);
     curl_mprintf("%s\n", dnsentry);
     slist2 = curl_slist_append(slist, dnsentry);
     if(!slist2) {
@@ -70,7 +70,7 @@ static CURLcode test_lib1506(char *URL)
     /* specify target */
     curl_msnprintf(target_url, sizeof(target_url),
                    "http://server%d.example.com:%s/path/1506%04i",
-                   i + 1, port, i + 1);
+                   (int)i + 1, port, (int)i + 1);
     target_url[sizeof(target_url) - 1] = '\0';
     easy_setopt(curl[i], CURLOPT_URL, target_url);
     /* go verbose */
