+++ tests/libtest/lib2502.c
@@ -34,7 +34,7 @@ static CURLcode test_lib2502(char *URL)
   CURL *curl[NUM_HANDLES] = {0};
   int running;
   CURLM *m = NULL;
-  int i;
+  size_t i;
   char target_url[256];
   char dnsentry[256];
   struct curl_slist *slist = NULL;
@@ -67,7 +67,7 @@ static CURLcode test_lib2502(char *URL)
     /* specify target */
     curl_msnprintf(target_url, sizeof(target_url),
                    "https://localhost:%s/path/2502%04i",
-                   port, i + 1);
+                   port, (int)i + 1);
     target_url[sizeof(target_url) - 1] = '\0';
     easy_setopt(curl[i], CURLOPT_URL, target_url);
     /* go http2 */
