+++ tests/libtest/lib540.c
@@ -163,7 +163,7 @@ static CURLcode loop(int num, CURLM *cm, const char *url, const char *userpwd,
       if(!msg)
         break;
       if(msg->msg == CURLMSG_DONE) {
-        int i;
+        size_t i;
         CURL *e = msg->easy_handle;
         curl_mfprintf(stderr, "R: %d - %s\n", (int)msg->data.result,
                 curl_easy_strerror(msg->data.result));
@@ -194,7 +194,7 @@ static CURLcode test_lib540(char *URL)
   struct curl_slist *headers = NULL;
   char buffer[246]; /* naively fixed-size */
   CURLcode res = CURLE_OK;
-  int i;
+  size_t i;
 
   for(i = 0; i < CURL_ARRAYSIZE(testeh); i++)
     testeh[i] = NULL;
